+++ src/fio/FioEngine.ts
@@ -2045,34 +2045,26 @@ export const parseAction = ({
       if (data.payee_public_key === publicKey) {
         ourReceiveAddresses.push(publicKey)
         if (data.actor === actor) {
-          nativeAmount = '0' // Self-transfer should not affect balance
           networkFee = dataMaxFee
+          nativeAmount = `-${networkFee}`
         } else {
           nativeAmount = dataAmount // Receiving funds
           networkFee = '0'
         }
       } else if (data.actor === actor) {
-        nativeAmount = `-${add(dataAmount, dataMaxFee)}` // Sending funds
         networkFee = dataMaxFee
+        nativeAmount = `-${add(dataAmount, networkFee)}` // Sending funds
       } else {
         // This action doesn't involve our account, so we should ignore it
         return { blockNum: action.block_num }
-      }
-
-      if (existingTx != null) {
-        // If we've already processed this transaction, we should ignore this action
-        if (existingTx.otherParams?.meta?.isTransferProcessed != null) {
-          return { blockNum: action.block_num }
-        }
-        // If this is a new action for an existing transaction, we should update the existing transaction
-        nativeAmount = add(existingTx.nativeAmount, nativeAmount)
       }
 
       otherParams.meta.isTransferProcessed = true
       break
 
     case 'stakefio': {
-      nativeAmount = `-${dataAmount}`
+      networkFee = dataMaxFee
+      nativeAmount = `-${networkFee}`
       otherParams.meta.isTransferProcessed = true
       break
     }
@@ -2084,11 +2076,11 @@ export const parseAction = ({
         txId: action.action_trace.trx_id,
         txName: actName
       }
-
-      // Unstake actions should have a corresponding reward 'transfer' action
-      // that was parsed right before this action, which only reports the reward
-      // portion of the unstake.
+      networkFee = dataMaxFee
       if (existingTx != null) {
+        // Unstake actions should have a corresponding reward 'transfer' action
+        // that was parsed right before this action, which only reports the reward
+        // portion of the unstake.
         otherParams = {
           ...existingTx.otherParams,
           ...otherParams,
@@ -2101,10 +2093,10 @@ export const parseAction = ({
             ...otherParams.meta
           }
         }
+        nativeAmount = sub(existingTx.nativeAmount, networkFee)
+      } else {
+        nativeAmount = `-${networkFee}`
       }
-      // Add the additional unstake amount that was requested by these actions
-      // to the already saved reward amount, if any.
-      nativeAmount = add(dataAmount, existingTx?.nativeAmount ?? '0')
 
       otherParams.meta.isTransferProcessed = true
       break
@@ -2126,9 +2118,6 @@ export const parseAction = ({
     // Fee or (unstake) reward transaction
     case 'transfer':
       {
-        const isRecipient = data.to === actor
-        nativeAmount = isRecipient ? `${fioAmount}` : `-${fioAmount}`
-
         // Some transfers might be rewards/yield from unstaking
         const isUnstakeRewardTx =
           otherParams.data != null &&
@@ -2141,8 +2130,11 @@ export const parseAction = ({
             txName: actName
           }
           networkFee = '0'
+          nativeAmount = fioAmount
         } else {
-          networkFee = isRecipient ? `-${fioAmount}` : fioAmount
+          const isRecipient = data.to === actor
+          networkFee = isRecipient ? `0` : fioAmount
+          nativeAmount = isRecipient ? '0' : `-${networkFee}`
         }
 
         if (existingTx != null) {
@@ -2162,6 +2154,7 @@ export const parseAction = ({
 
           if (otherParams.meta.isTransferProcessed != null) {
             if (data.to !== actor) {
+              networkFee = dataMaxFee
               nativeAmount = sub(existingTx.nativeAmount, networkFee)
             } else {
               networkFee = '0'