+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Unreleased
 
 - fixed: (EVM) Fixed ETH requires a resync or log back in to see new transaction
-
+- fixed: (FIO) Fix amount handling in parseActions
+  
 ## 4.24.2 (2024-09-12)
 
 - fixed: (EVM) Merge duplicate token transaction data from evmscan fixing incorrect transaction native amounts